requirements.txt
couchbase==4.3.4
python-dotenv==1.0.1
